requirements.txt
numpy
torch
mlflow
bonsai-nn-library
tqdm
pandas
seaborn
jsonargparse